moved outcome.show() to canvas.get_outcome()
